refactor(api): Move old [id] files under [id] folder

+++ server/api/schedule/[id]/index.delete.ts
@@ -6,5 +6,10 @@ export default defineEventHandler(async (event) => {
 
   const {id} = await getValidatedRouterParams(event, v.parser(idParamSchema));
 
+  const scheduleItem = await prisma.scheduleItem.findUnique({where: {id}});
+  if (!scheduleItem) {
+    throw createError({statusCode: 404, message: "Partie de planning introuvable"});
+  }
+
   return prisma.scheduleItem.delete({where: {id}});
 });
